admin_panel/pyproject.toml
[tool.poetry]
name = "admin_panel"
version = "0.0.1"
description = ""
authors = ["T23 <petr777@yahoo.com>"]

[tool.poetry.dependencies]
# General
python = "^3.8"
python-dotenv = "^0.19.2"
flake8 = "^4.0.1"
loguru = "^0.6.0"
psycopg2-binary = "^2.9.3"

# django
django = "^3.2"
django-split-settings = "^1.1.0"
django-debug-toolbar = "^3.2.4"
gunicorn='20.0.4'

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
